Arquitetura-Medalhao/pyproject.toml
[tool.poetry]
name = "video1"
version = "0.1.0"
description = ""
authors = ["Bruno-Gomes-QA <bruno.gomes.qa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kagglehub = "^0.3.3"
boto3 = "^1.35.49"
python-dotenv = "^1.0.1"
blue = "^0.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
